checkpoint: renames "URLComponent/URLOrigin.ts" to "URLComponent/Origin.ts"
